+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All Notable changes to `laravel-imap` will be documented in this file
 
-## 2015-08-21
+## 1.0.2 (2015-08-22)
 
-### First commit
+Bugfixes:
+
+- Fix charset issues in message body
+
+## 1.0.1 (2015-08-22)
+
+Bugfixes:
+
+ - composer.json now works
+
+## 1.0.0 (2015-08-21)
+
+First commit